# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Proof-Asset Registry (PAR) API",
    "version": "1.0.0",
    "description": "Privacy-first, enterprise-grade platform for secure registration, verification, and lifecycle management of cryptographic proofs. Supports zero-knowledge proofs, JSON Web Signatures, W3C-compliant credential status tracking, and append-only audit transparency.",
    "contact": {
      "name": "PAR Support",
      "url": "https://github.com/your-org/proof-asset-registry"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.yourproofregistry.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/proof-assets": {
      "get": {
        "tags": ["Proof Assets"],
        "summary": "List all proof assets",
        "description": "Retrieves all proof assets in the registry",
        "operationId": "listProofAssets",
        "security": [{ "ApiKeyAuth": [] }],
        "responses": {
          "200": {
            "description": "List of proof assets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ProofAsset" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Proof Assets"],
        "summary": "Register a new proof asset",
        "description": "Registers a new cryptographic proof asset with verification and receipt generation",
        "operationId": "registerProofAsset",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProofAssetCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Proof asset registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "proofAssetId": { "type": "string" },
                    "commitment": { "type": "string" },
                    "receipt": { "type": "string" },
                    "statusRef": {
                      "type": "object",
                      "properties": {
                        "statusListUrl": { "type": "string" },
                        "statusListIndex": { "type": "string" },
                        "statusPurpose": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key authentication for partners. Format: `{keyId}.{secret}` or `ApiKey {keyId}.{secret}` in Authorization header."
      },
      "AdminApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Admin API key authentication. Requires `admin:*` scope. Format: `ApiKey {keyId}.{secret}`"
      }
    },
    "schemas": {
      "ProofAsset": {
        "type": "object",
        "properties": {
          "proofAssetId": { "type": "string" },
          "partnerId": { "type": "string" },
          "issuerDid": { "type": "string" },
          "verifierDid": { "type": "string" },
          "proofFormat": { "type": "string", "enum": ["jws", "zk-snark", "bbs+"] },
          "proofDigest": { "type": "string" },
          "digestAlg": { "type": "string" },
          "policyHash": { "type": "string" },
          "constraintHash": { "type": "string" },
          "commitment": { "type": "string" },
          "statusListUrl": { "type": "string" },
          "statusListIndex": { "type": "string" },
          "statusPurpose": { "type": "string" },
          "verificationStatus": { "type": "string" },
          "verifierProofRef": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "verifiedAt": { "type": "string", "format": "date-time" }
        }
      },
      "ProofAssetCreate": {
        "type": "object",
        "required": ["issuer_did", "verifier_did", "proof_format", "proof_digest", "digest_alg", "policy_hash", "constraint_hash"],
        "properties": {
          "issuer_did": { "type": "string", "description": "DID of proof issuer" },
          "verifier_did": { "type": "string", "description": "DID of verifier" },
          "proof_format": { "type": "string", "enum": ["jws", "zk-snark", "bbs+"] },
          "proof_digest": { "type": "string", "description": "Cryptographic digest of proof bytes" },
          "digest_alg": { "type": "string", "enum": ["sha2-256", "sha3-256", "blake3", "multihash"] },
          "policy_hash": { "type": "string", "description": "Hash of verification policy" },
          "constraint_hash": { "type": "string", "description": "Hash of verification constraints" },
          "proof_uri": { "type": "string", "description": "URI to fetch proof for verification" },
          "proof_bytes": { "type": "string", "format": "byte", "description": "Base64-encoded proof bytes" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "detail": { "type": "string" }
        }
      },
...
    }
  }
}